create module for align function
