+++ src/app.js
@@ -3,10 +3,12 @@ import createTable from './js/table';
 import resizeCells from './js/cellResizing';
 import saveContent from './js/textSaving';
 import createIcons from './js/icons';
-import selectCells from './js/selectingCells';
+import { selectCells } from './js/selectingCells';
+import toAlign from './js/formattingText';
 
 createTable('#table', 20, 10);
 createIcons();
 resizeCells();
 saveContent();
 selectCells();
+toAlign();
