+++ src/js/selectingCells.js
@@ -1,7 +1,5 @@
-import { saveState } from './localStorage';
-
-export default function selectCells() {
-  let cells;
+let cells;
+export function selectCells() {
   const table = document.getElementById('table');
   function cancelSelection() {
     cells.forEach(c => document.getElementById(`${c}`).classList.remove('table__selected'));
@@ -43,16 +41,9 @@ export default function selectCells() {
       eve.target.classList.add('table__selected');
     };
   };
+}
 
-  document.onclick = function toAlign(event) {
-    if (event.target.hasAttribute('data-button')) {
-      const pos = event.target.getAttribute('data-button');
-      cells.forEach(c => {
-        const textCell = document.getElementById(`${c}`);
-        textCell.style.textAlign = `${pos}`;
-        textCell.setAttribute('data-align', `${pos}`);
-        saveState('text-align', c, pos);
-      });
-    }
-  };
+export function getSelectCells() {
+  const cellsAr = cells;
+  return cellsAr;
 }
